+++ src/pyinterp/tests/test_binning.py
@@ -36,11 +36,11 @@ def build_binning2d_instance(dtype):
     linear_mean = binning.variable('mean')
     assert isinstance(simple_mean, np.ndarray)
     assert np.any(linear_mean != simple_mean)
-    assert isinstance(binning.variable("sum"), np.ndarray)
-    assert isinstance(binning.variable("count"), np.ndarray)
+    assert isinstance(binning.variable('sum'), np.ndarray)
+    assert isinstance(binning.variable('count'), np.ndarray)
 
     with pytest.raises(ValueError):
-        binning.variable("_")
+        binning.variable('_')
 
 
 def test_binning2d():
@@ -56,22 +56,22 @@ def test_binning2d_dask():
     y_axis = Axis(np.linspace(-80, 80, 1))
     binning = Binning2D(x_axis, y_axis)
 
-    x = da.full((4096 * 8, ), -180.0, dtype="f8", chunks=4096)
-    y = da.full((4096 * 8, ), -80.0, dtype="f8", chunks=4096)
+    x = da.full((4096 * 8, ), -180.0, dtype='f8', chunks=4096)
+    y = da.full((4096 * 8, ), -80.0, dtype='f8', chunks=4096)
     z = da.random.uniform(size=4096 * 8, chunks=4096)
 
     binning = binning.push_delayed(x, y, z).compute()
 
-    assert np.all(binning.variable("count") == 32768)
-    assert binning.variable("sum_of_weights")[0, 0] == 32768
-    assert binning.variable("mean")[0, 0] == pytest.approx(z.mean().compute())
-    assert binning.variable("variance")[0, 0] == pytest.approx(
+    assert np.all(binning.variable('count') == 32768)
+    assert binning.variable('sum_of_weights')[0, 0] == 32768
+    assert binning.variable('mean')[0, 0] == pytest.approx(z.mean().compute())
+    assert binning.variable('variance')[0, 0] == pytest.approx(
         z.std().compute()**2)  # type: ignore
-    assert binning.variable("sum")[0, 0] == pytest.approx(z.sum().compute())
+    assert binning.variable('sum')[0, 0] == pytest.approx(z.sum().compute())
     if HAVE_SCIPY:
-        assert binning.variable("kurtosis")[0, 0] == pytest.approx(
+        assert binning.variable('kurtosis')[0, 0] == pytest.approx(
             das.kurtosis(z).compute())  # type: ignore
-        assert binning.variable("skewness")[0, 0] == pytest.approx(
+        assert binning.variable('skewness')[0, 0] == pytest.approx(
             das.skew(z).compute())  # type: ignore
 
 
@@ -91,11 +91,11 @@ def build_binning1d_instance(dtype):
     binning.push(lon, ds.mss)
     mean = binning.variable('mean')
     assert isinstance(mean, np.ndarray)
-    assert isinstance(binning.variable("sum"), np.ndarray)
-    assert isinstance(binning.variable("count"), np.ndarray)
+    assert isinstance(binning.variable('sum'), np.ndarray)
+    assert isinstance(binning.variable('count'), np.ndarray)
 
     with pytest.raises(ValueError):
-        binning.variable("_")
+        binning.variable('_')
 
 
 def test_binning1d():
@@ -110,19 +110,19 @@ def test_binning1d_dask():
     x_axis = Axis(np.linspace(-180, 180, 1), is_circle=True)
     binning = Binning1D(x_axis)
 
-    x = da.full((4096 * 8, ), -180.0, dtype="f8", chunks=4096)
+    x = da.full((4096 * 8, ), -180.0, dtype='f8', chunks=4096)
     z = da.random.uniform(size=4096 * 8, chunks=4096)
 
     binning = binning.push_delayed(x, z).compute()
 
-    assert np.all(binning.variable("count") == 32768)
-    assert binning.variable("sum_of_weights")[0, 0] == 32768
-    assert binning.variable("mean")[0, 0] == pytest.approx(z.mean().compute())
-    assert binning.variable("variance")[0, 0] == pytest.approx(
+    assert np.all(binning.variable('count') == 32768)
+    assert binning.variable('sum_of_weights')[0, 0] == 32768
+    assert binning.variable('mean')[0, 0] == pytest.approx(z.mean().compute())
+    assert binning.variable('variance')[0, 0] == pytest.approx(
         z.std().compute()**2)  # type: ignore
-    assert binning.variable("sum")[0, 0] == pytest.approx(z.sum().compute())
+    assert binning.variable('sum')[0, 0] == pytest.approx(z.sum().compute())
     if HAVE_SCIPY:
-        assert binning.variable("kurtosis")[0, 0] == pytest.approx(
+        assert binning.variable('kurtosis')[0, 0] == pytest.approx(
             das.kurtosis(z).compute())  # type: ignore
-        assert binning.variable("skewness")[0, 0] == pytest.approx(
+        assert binning.variable('skewness')[0, 0] == pytest.approx(
             das.skew(z).compute())  # type: ignore
